Implement tangent free normal mapping
