+++ src/Game.cpp
@@ -195,7 +195,7 @@ void Game::CreateDeviceDependentResources()
 
     std::vector<uint8_t> blob = DX::ReadData(L"skinning.cso");
     DX::ThrowIfFailed(m_renderer.getDevice()->CreateComputeShader(blob.data(), blob.size(), nullptr, m_renderer.computeShader.ReleaseAndGetAddressOf()));
-    bdr::gltf::SceneData sceneData{ &m_scene, &m_renderer, "FlightHelmetV2/", "FlightHelmet.gltf" };
+    bdr::gltf::SceneData sceneData{ &m_scene, &m_renderer, "PollyV2/", "project_polly.gltf" };
     bdr::gltf::loadModel(sceneData);
 }
 
@@ -216,8 +216,8 @@ void Game::CreateWindowSizeDependentResources()
     baseView.scene = &m_scene;
     baseView.setCamera(&m_camera);
 
-    bdr::addBasicPass(m_renderGraph, &baseView);
     bdr::addSkinningPass(m_renderGraph, &baseView);
+    bdr::addBasicPass(m_renderGraph, &baseView);
 }
 
 void Game::OnDeviceLost()
