+++ src/Shaders/basic_vs.hlsl
@@ -17,7 +17,6 @@ struct VSInput
     float3 Position : SV_Position;
     float3 Normal : NORMAL;
     float2 UV : TEXCOORD;
-    float4 Tangent : TANGENT;
 };
 
 struct VSOutput
@@ -25,8 +24,6 @@ struct VSOutput
     float4 PositionCS : SV_Position;
     float3 PositionWS : POSITIONWS;
     float3 NormalWS : NORMALWS;
-    float3 TangentWS : TANGENTWS;
-    float3 BitangentWS : BITANGENTWS;
     float2 vUV : TEXCOORD;
 };
 
@@ -37,9 +34,7 @@ VSOutput main(in VSInput input)
     output.PositionCS = mul(float4(output.PositionWS, 1.0f), VP);
     output.vUV = input.UV;
     
-    output.NormalWS = normalize(mul(float4(input.Normal, 1.0f), invModel).xyz);
-    output.TangentWS = normalize(mul(float4(input.Tangent.xyz, 0.0), model).xyz);
-    output.BitangentWS = normalize(cross(output.NormalWS, output.TangentWS)) * input.Tangent.w;
+    output.NormalWS = mul(float4(input.Normal, 1.0f), invModel).xyz;
 
     return output;
 }
