+++ src/skinning.hlsl
@@ -9,11 +9,9 @@ Buffer<float3> in_POS : register(t1);
 Buffer<float3> in_NORM : register(t2);
 Buffer<uint4> in_INDICES : register(t3);
 Buffer<float4> in_WEIGHTS : register(t4);
-Buffer<float4> in_TANG : register(t5);
 
 RWByteAddressBuffer out_Pos : register(u0);
 RWByteAddressBuffer out_Norm : register(u1);
-RWByteAddressBuffer out_Tan : register(u2);
 
 [numthreads(64, 1, 1)]
 void main(uint3 DTid : SV_DispatchThreadID)
@@ -27,7 +25,6 @@ void main(uint3 DTid : SV_DispatchThreadID)
     {
         float3 position = in_POS.Load(DTid.x);
         float3 normal = in_NORM.Load(DTid.x);
-        float4 tangent = in_TANG.Load(DTid.x);
         
         uint4 joints = in_INDICES.Load(DTid.x);
         float4 weights = (in_WEIGHTS.Load(DTid.x));
@@ -40,10 +37,8 @@ void main(uint3 DTid : SV_DispatchThreadID)
         float4 newPos = mul(float4(position, 1.0), skinMatrix);
         position = newPos.xyz / newPos.w;
         float3 newNormal = normalize(mul(normal, normalMatrix));
-        float4 newTangent = mul(float4(tangent.xyz, 0.0), skinMatrix);
 
         out_Pos.Store3(posOffset, asuint(position));
         out_Norm.Store3(posOffset, asuint(newNormal));
-        out_Tan.Store4(tanOffset, asuint(float4(newTangent.xyz, tangent.w)));
     }
 }
